Done. Could use Processment improvment + tests

+++ src/hopeless/logic/BoardComparator.java
@@ -6,6 +6,6 @@ public class BoardComparator implements Comparator<Board>{
 
 	@Override
 	public int compare(Board b1, Board b2){
-		return Integer.compare(b1.finalCost, b2.finalCost);
+		return Double.compare(b1.finalCost, b2.finalCost);
 	}
 }
